smart-router/requirements.txt
# smart-router dependencies
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
anthropic==0.34.2
openai==1.51.0
httpx==0.27.2
pytest==8.3.3
streamlit==1.39.0
groq==0.11.0